+++ .eslintrc.js
@@ -23,6 +23,7 @@ module.exports = {
     rules: {
       'jsx-a11y/label-has-for': 'off',
       'jsx-a11y/no-noninteractive-tabindex': 'off',
+      'max-len': 'off',
     },
   }, {
     files: [