+++ src/hooks/use-mounted.js
@@ -1,0 +1,18 @@
+import {
+  useEffect,
+  useRef,
+} from 'react';
+
+export default function useMounted() {
+  const mounted = useRef(false);
+
+  useEffect(() => {
+    mounted.current = true;
+
+    return () => {
+      mounted.current = false;
+    };
+  }, []);
+
+  return mounted;
+}